Refactor to require relative per requirements for spec harness all tests pass
