+++ lib/stat_tracker.rb
