+++ spec/spec_helper.rb
@@ -7,4 +7,4 @@ require './lib/team'
 require './lib/game_statistics'
 require './lib/league_statistics'
 require './lib/season_statistics'
-require './lib/StatTracker.rb'
+require './lib/stat_tracker.rb'
